+++ src/main/java/tornadofx/FX.kt
@@ -9,7 +9,7 @@ class FX {
         lateinit var primaryStage: Stage
         lateinit var application: App
         val stylesheets = ArrayList<String>()
-        val components = HashMap<KClass<out Component>, Component>()
+        val components = HashMap<KClass<out Injectable>, Injectable>()
         val lock = Any()
     }
 }
@@ -19,13 +19,10 @@ fun importStylesheet(stylesheet: String) {
     FX.stylesheets.add(css.toExternalForm())
 }
 
-inline fun <reified T : Component> find(): T = find(T::class)
+inline fun <reified T : Injectable> find(): T = find(T::class)
 
 @Suppress("UNCHECKED_CAST")
-fun <T : Component> find(type: KClass<T>): T {
-    if (Fragment::class.java.isAssignableFrom(type.java))
-        return type.java.newInstance()
-
+fun <T : Injectable> find(type: KClass<T>): T {
     if (!FX.components.containsKey(type)) {
         synchronized(FX.lock) {
             if (!FX.components.containsKey(type))